Add Marshal/Unmarshal for HLL++
This is to make it easy to persist your HLLs to disk/database.

When marshaling non-sparse HLL, it saves space by using less than 6 bits per
register if possible (e.g. if no registers are greater than 31, you can use 5
bits per register instead of 6).

The marshal format also includes a version number, so backwards compatibility
can be maintained in the future if the format is changed.